feat!: migrate scoring APIs to double precision for v6.0.0

+++ FuzzySharp.Benchmarks/ExtractSelectionBenchmarks.cs
@@ -203,12 +203,12 @@ public class ExtractSelectionOneBenchmarks
 
     private sealed class CheapScorer : IRatioScorer
     {
-        public int Score(string input1, string input2)
+        public double Score(string input1, string input2)
         {
             return ScoreFromChoice(input2);
         }
 
-        public int Score(string input1, string input2, Func<string, string> preprocessor)
+        public double Score(string input1, string input2, Func<string, string> preprocessor)
         {
             return Score(preprocessor(input1), preprocessor(input2));
         }
@@ -216,7 +216,7 @@ public class ExtractSelectionOneBenchmarks
 
     private sealed class CheapCachedScorer : ICachedRatioScorer
     {
-        public int Score(string input2)
+        public double Score(string input2)
         {
             return ScoreFromChoice(input2);
         }
@@ -335,12 +335,12 @@ public class ExtractSelectionTopBenchmarks
 
     private sealed class CheapScorer : IRatioScorer
     {
-        public int Score(string input1, string input2)
+        public double Score(string input1, string input2)
         {
             return ScoreFromChoice(input2);
         }
 
-        public int Score(string input1, string input2, Func<string, string> preprocessor)
+        public double Score(string input1, string input2, Func<string, string> preprocessor)
         {
             return Score(preprocessor(input1), preprocessor(input2));
         }
@@ -348,7 +348,7 @@ public class ExtractSelectionTopBenchmarks
 
     private sealed class CheapCachedScorer : ICachedRatioScorer
     {
-        public int Score(string input2)
+        public double Score(string input2)
         {
             return ScoreFromChoice(input2);
         }
@@ -497,12 +497,12 @@ public class ExtractSelectionFocusedBenchmarks
 
     private sealed class CheapScorer : IRatioScorer
     {
-        public int Score(string input1, string input2)
+        public double Score(string input1, string input2)
         {
             return ScoreFromChoice(input2);
         }
 
-        public int Score(string input1, string input2, Func<string, string> preprocessor)
+        public double Score(string input1, string input2, Func<string, string> preprocessor)
         {
             return Score(preprocessor(input1), preprocessor(input2));
         }
@@ -510,7 +510,7 @@ public class ExtractSelectionFocusedBenchmarks
 
     private sealed class CheapCachedScorer : ICachedRatioScorer
     {
-        public int Score(string input2)
+        public double Score(string input2)
         {
             return ScoreFromChoice(input2);
         }
